\name{df1}
\alias{df1}
\docType{data}
\title{
Hox pupil popularity data with missing popularity scores
}
\description{
df1 data set
}
\usage{data("df1")}
\format{
  A data frame with 2000 observations on the following 7 variables.
  \describe{
    \item{\code{PUPIL}}{a numeric vector}
    \item{\code{SCHOOL}}{a numeric vector}
    \item{\code{POPULAR}}{a numeric vector}
    \item{\code{SEX}}{a numeric vector}
    \item{\code{TEXP}}{a numeric vector}
    \item{\code{CONST}}{a numeric vector}
    \item{\code{TEACHPOP}}{a numeric vector}
  }
}
\details{
The original, complete dataset was generated by Joop Hox as an example of well-behaved multilevel data set. The distributed data contains missing data in pupil popularity.

}
\source{
The Heart failure data set comes from the R package"mice".
}
\references{
Hox, J. J. (2002) Multilevel analysis. Techniques and applications. Mahwah, NJ: Lawrence Erlbaum.
}
\examples{
data(df1)
## maybe str(df1) ; plot(df1) ...
}
\keyword{datasets}
